[chore] 버전 조회 엔드포인트 수정
